refactor(ui): introduce svgr transformer
use svgr transformer to translate svg files into react component direactly
