+++ packages/jest-config/package.json
@@ -27,6 +27,7 @@
     "@types/node": "16",
     "eslint": "^8.10.0",
     "jest": "^27.5.1",
+    "jest-transformer-svg": "^1.0.2",
     "lint-staged": "^12.0.0",
     "prettier": "^2.3.2",
     "ts-node": "^10.0.0",
