+++ packages/ui/.parcelrc
@@ -1,0 +1,8 @@
+{
+  "extends": "@parcel/config-default",
+  "transformers": {
+    "*.svg": [
+      "@parcel/transformer-svg-react"
+    ]
+  }
+}
