+++ packages/ui/package.json
@@ -21,6 +21,7 @@
     "@logto/schemas": "^0.1.0",
     "@parcel/core": "^2.5.0",
     "@parcel/transformer-sass": "^2.5.0",
+    "@parcel/transformer-svg-react": "^2.5.0",
     "@peculiar/webcrypto": "^1.3.3",
     "@silverhand/eslint-config": "^0.12.0",
     "@silverhand/eslint-config-react": "^0.11.0",
@@ -56,7 +57,9 @@
     "react-string-replace": "^1.0.0",
     "react-timer-hook": "^3.0.5",
     "stylelint": "^13.13.1",
-    "typescript": "^4.6.2"
+    "typescript": "^4.6.2",
+    "react-device-detect": "^2.2.2",
+    "use-debounced-loader": "^0.1.1"
   },
   "alias": {
     "@/*": "./src/$1"
@@ -80,9 +83,5 @@
       ]
     }
   },
-  "prettier": "@silverhand/eslint-config/.prettierrc",
-  "dependencies": {
-    "react-device-detect": "^2.2.2",
-    "use-debounced-loader": "^0.1.1"
-  }
+  "prettier": "@silverhand/eslint-config/.prettierrc"
 }
