+++ packages/ui/src/assets/index.d.ts
@@ -1,0 +1,6 @@
+interface SvgComponent extends React.StatelessComponent<React.SVGAttributes<SVGElement>> {}
+
+declare module '*.svg' {
+  const value: SvgComponent;
+  export default value;
+}
