+++ packages/ui/src/components/ConfirmModal/Acmodal.module.scss
@@ -25,6 +25,12 @@
   @include _.flex-row;
   justify-content: space-between;
   margin-bottom: _.unit(4);
+
+  > svg {
+    color: var(--color-icon);
+    width: 24px;
+    height: 24px;
+  }
 }
 
 .content {
