--- packages/ui/src/components/Icons/index.ts
@@ -1,9 +1,0 @@
-export { default as CloseIcon } from './CloseIcon';
-export { default as ClearIcon } from './ClearIcon';
-export { default as PrivacyIcon } from './PrivacyIcon';
-export { default as DownArrowIcon } from './DownArrowIcon';
-export { default as LoadingIcon } from './LoadingIcon';
-export { default as NavArrowIcon } from './NavArrowIcon';
-export { default as ExpandMoreIcon } from './ExpandMoreIcon';
-export { default as ErrorIcon } from './ErrorIcon';
-export { default as CheckBoxIcon } from './CheckBoxIcon';
