+++ packages/ui/src/components/LoadingLayer/index.module.scss
@@ -19,7 +19,7 @@
 }
 
 .loadingIcon {
-  animation: rotating 1.5s ease-in-out infinite;
+  animation: rotating 1s linear infinite;
 }
 
 @keyframes rotating {
