+++ packages/ui/src/containers/SocialSignIn/PrimarySocialSignIn.module.scss
@@ -9,6 +9,7 @@
 }
 
 .expandIcon {
+  color: var(--color-icon);
   width: 20px;
   height: 20px;
